*** empty log message ***

+++ libEM/util.h
@@ -145,7 +145,7 @@ namespace EMAN {
 
 	static inline float angle_sub_2pi(float x, float y)
 	{
-	    float r = fmod(fabs(x-y), 2 * M_PI);
+	    float r = fmod(fabs(x-y), (float)(2 * M_PI));
 	    if (r > M_PI) {
 		r = 2.0 * M_PI - r;
 	    }
@@ -155,7 +155,7 @@ namespace EMAN {
 
 	static inline float angle_sub_pi(float x, float y)
 	{
-	    float r = fmod(fabs(x-y), M_PI);
+	    float r = fmod(fabs(x-y), (float)M_PI);
 	    if (r > M_PI / 2.0) {
 		r = M_PI - r;
 	    }
